feat: add test_views
